+++ js/main.js
@@ -1,5 +1,5 @@
 import './data.js';
 import './markup.js';
-import {makeUnactive} from './makeActive_Unactive.js';
+import {makeUnactive} from './switchingActivity.js';
 
 makeUnactive();
