+++ js/switchingActivity.js
